requirements.txt
requests
pytest
kafka-python==2.0.2
